Simplify ClusterChangeHandler
