+++ ambry-clustermap/src/main/java/com/github/ambry/clustermap/DatacenterInitializer.java
@@ -17,14 +17,9 @@ package com.github.ambry.clustermap;
 
 import com.github.ambry.config.ClusterMapConfig;
 import com.github.ambry.utils.Utils;
-import java.nio.ByteBuffer;
-import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.atomic.AtomicLong;
 import org.apache.helix.HelixManager;
 import org.apache.helix.InstanceType;
 import org.apache.helix.PropertyType;
@@ -46,15 +41,10 @@ class DatacenterInitializer {
   private final HelixFactory helixFactory;
   private final ClusterMapUtils.DcZkInfo dcZkInfo;
   private final String dcName;
-
-  // Fields to pass into Dynamic ClusterChangeHandler
+  private final HelixClusterManager helixClusterManager;
   private final String selfInstanceName;
-  private final Map<String, Map<String, String>> partitionOverrideInfoMap;
-  private final HelixClusterManager.ClusterChangeHandlerCallback clusterChangeHandlerCallback;
   private final HelixClusterManager.HelixClusterManagerCallback helixClusterManagerCallback;
-  private final HelixClusterManagerMetrics helixClusterManagerMetrics;
   private final DataNodeConfigSourceMetrics dataNodeConfigSourceMetrics;
-  private final AtomicLong sealedStateChangeCounter;
 
   /**
    * @param clusterMapConfig {@link ClusterMapConfig} to help some admin operations
@@ -62,37 +52,23 @@ class DatacenterInitializer {
    * @param helixFactory the {@link HelixFactory} instance to construct managers.
    * @param dcZkInfo info about the DC, like connection string, name, and replica type
    * @param selfInstanceName the name of instance on which {@link HelixClusterManager} resides.
-   * @param partitionOverrideInfoMap a map specifying partitions whose state should be overridden.
-   * @param clusterChangeHandlerCallback a call back that allows current handler to update cluster-wide info.
    * @param helixClusterManagerCallback a help class to get cluster state from all DCs.
-   * @param helixClusterManagerMetrics metrics that help track of cluster changes and infos.
    * @param dataNodeConfigSourceMetrics metrics related to {@link DataNodeConfigSource}.
-   * @param sealedStateChangeCounter a counter that records event when replica is sealed or unsealed
-   * @param partitionMap a map from serialized bytes to corresponding partition.
-   * @param partitionNameToAmbryPartition a map from partition name to {@link AmbryPartition} object.
-   * @param ambryPartitionToAmbryReplicas a map from {@link AmbryPartition} to its replicas.
+   * @param helixClusterManager
    */
   DatacenterInitializer(ClusterMapConfig clusterMapConfig, HelixManager localManager, HelixFactory helixFactory,
       ClusterMapUtils.DcZkInfo dcZkInfo, String selfInstanceName,
-      Map<String, Map<String, String>> partitionOverrideInfoMap,
-      HelixClusterManager.ClusterChangeHandlerCallback clusterChangeHandlerCallback,
       HelixClusterManager.HelixClusterManagerCallback helixClusterManagerCallback,
-      HelixClusterManagerMetrics helixClusterManagerMetrics, DataNodeConfigSourceMetrics dataNodeConfigSourceMetrics,
-      AtomicLong sealedStateChangeCounter, ConcurrentHashMap<ByteBuffer, AmbryPartition> partitionMap,
-      ConcurrentHashMap<String, AmbryPartition> partitionNameToAmbryPartition,
-      ConcurrentHashMap<AmbryPartition, Set<AmbryReplica>> ambryPartitionToAmbryReplicas) {
+      DataNodeConfigSourceMetrics dataNodeConfigSourceMetrics, HelixClusterManager helixClusterManager) {
     this.clusterMapConfig = clusterMapConfig;
     this.localManager = localManager;
     this.helixFactory = helixFactory;
     this.dcZkInfo = dcZkInfo;
     this.selfInstanceName = selfInstanceName;
-    this.partitionOverrideInfoMap = partitionOverrideInfoMap;
-    this.clusterChangeHandlerCallback = clusterChangeHandlerCallback;
     this.helixClusterManagerCallback = helixClusterManagerCallback;
-    this.helixClusterManagerMetrics = helixClusterManagerMetrics;
     this.dataNodeConfigSourceMetrics = dataNodeConfigSourceMetrics;
-    this.sealedStateChangeCounter = sealedStateChangeCounter;
     dcName = dcZkInfo.getDcName();
+    this.helixClusterManager = helixClusterManager;
   }
 
   /**
@@ -159,10 +135,9 @@ class DatacenterInitializer {
       manager = helixFactory.getZkHelixManagerAndConnect(clusterMapConfig.clusterMapClusterName, selfInstanceName,
           InstanceType.SPECTATOR, zkConnectStr);
     }
-    HelixClusterChangeHandler clusterChangeHandler =
-        new DynamicClusterChangeHandler(clusterMapConfig, dcName, selfInstanceName, partitionOverrideInfoMap,
-            helixClusterManagerCallback, clusterChangeHandlerCallback, helixClusterManagerMetrics,
-            this::onInitializationFailure, sealedStateChangeCounter);
+    HelixClusterManager.HelixClusterChangeHandler clusterChangeHandler =
+        helixClusterManager.new HelixClusterChangeHandler(dcName, helixClusterManagerCallback,
+            this::onInitializationFailure);
     // Create RoutingTableProvider of each DC to keep track of partition(replicas) state. Here, we use current
     // state based RoutingTableProvider to remove dependency on Helix's pipeline and reduce notification latency.
     logger.info("Creating routing table provider associated with Helix manager at {}", zkConnectStr);
@@ -193,7 +168,7 @@ class DatacenterInitializer {
     clusterChangeHandler.setRoutingTableSnapshot(routingTableProvider.getRoutingTableSnapshot());
     // the initial routing table change should populate the instanceConfigs. If it's empty that means initial
     // change didn't come and thread should wait on the init latch to ensure routing table snapshot is non-empty
-    if (clusterChangeHandler.getRoutingTableSnapshot().getInstanceConfigs().isEmpty()) {
+    if (helixClusterManagerCallback.getRoutingTableSnapshot(dcName).getInstanceConfigs().isEmpty()) {
       // Periodic refresh in routing table provider is enabled by default. In worst case, routerUpdater should
       // trigger routing table change within 5 minutes
       logger.info("Routing table snapshot in {} is currently empty. Waiting for initial notification.", dcName);
