+++ ambry-clustermap/src/main/java/com/github/ambry/clustermap/HelixClusterManager.java
@@ -15,6 +15,7 @@ package com.github.ambry.clustermap;
 
 import com.codahale.metrics.MetricRegistry;
 import com.github.ambry.config.ClusterMapConfig;
+import com.github.ambry.utils.Pair;
 import com.github.ambry.utils.SystemTime;
 import java.io.IOException;
 import java.io.InputStream;
@@ -30,10 +31,22 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Consumer;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import org.apache.helix.AccessOption;
 import org.apache.helix.HelixManager;
 import org.apache.helix.InstanceType;
+import org.apache.helix.NotificationContext;
+import org.apache.helix.api.listeners.IdealStateChangeListener;
+import org.apache.helix.api.listeners.LiveInstanceChangeListener;
+import org.apache.helix.api.listeners.RoutingTableChangeListener;
+import org.apache.helix.model.IdealState;
+import org.apache.helix.model.LiveInstance;
 import org.apache.helix.spectator.RoutingTableSnapshot;
 import org.apache.helix.store.zk.ZkHelixPropertyStore;
 import org.apache.helix.zookeeper.datamodel.ZNRecord;
@@ -64,11 +77,19 @@ public class HelixClusterManager implements ClusterMap {
   private final ConcurrentHashMap<AmbryPartition, Set<AmbryReplica>> ambryPartitionToAmbryReplicas =
       new ConcurrentHashMap<>();
   private final ConcurrentHashMap<ByteBuffer, AmbryPartition> partitionMap = new ConcurrentHashMap<>();
+  private final ConcurrentHashMap<String, Set<AmbryDataNode>> dcToNodes = new ConcurrentHashMap<>();
+  // A map whose key is ambry datanode and value is a map of partitionId to corresponding replica associated with this datanode
+  private final ConcurrentHashMap<AmbryDataNode, ConcurrentHashMap<String, AmbryReplica>> ambryDataNodeToAmbryReplicas =
+      new ConcurrentHashMap<>();
+  private final ConcurrentHashMap<AmbryDataNode, Set<AmbryDisk>> ambryDataNodeToAmbryDisks = new ConcurrentHashMap<>();
+  private final Map<String, ConcurrentHashMap<String, String>> partitionToResourceNameByDc = new ConcurrentHashMap<>();
+  private final Map<String, AtomicReference<RoutingTableSnapshot>> dcToRoutingTableSnapshotRef = new HashMap<>();
+  private final ConcurrentHashMap<String, AmbryDataNode> instanceNameToAmbryDataNode = new ConcurrentHashMap<>();
+  private final AtomicLong errorCount = new AtomicLong(0);
   private final AtomicLong clusterWideRawCapacityBytes = new AtomicLong(0);
   private final AtomicLong clusterWideAllocatedRawCapacityBytes = new AtomicLong(0);
   private final AtomicLong clusterWideAllocatedUsableCapacityBytes = new AtomicLong(0);
   private final HelixClusterManagerCallback helixClusterManagerCallback;
-  private final ClusterChangeHandlerCallback clusterChangeHandlerCallback;
   private final byte localDatacenterId;
   private final AtomicLong sealedStateChangeCounter = new AtomicLong(0);
   private final PartitionSelectionHelper partitionSelectionHelper;
@@ -98,7 +119,6 @@ public class HelixClusterManager implements ClusterMap {
     clusterName = clusterMapConfig.clusterMapClusterName;
     selfInstanceName = instanceName;
     helixClusterManagerCallback = new HelixClusterManagerCallback();
-    clusterChangeHandlerCallback = new ClusterChangeHandlerCallback();
     helixClusterManagerMetrics = new HelixClusterManagerMetrics(metricRegistry, helixClusterManagerCallback);
     Map<String, DcZkInfo> dataCenterToZkAddress = null;
     HelixManager localManager = null;
@@ -118,9 +138,7 @@ public class HelixClusterManager implements ClusterMap {
         // Initialize from every remote datacenter in a separate thread to speed things up.
         DatacenterInitializer initializer =
             new DatacenterInitializer(clusterMapConfig, localManager, helixFactory, dcZkInfo, selfInstanceName,
-                partitionOverrideInfoMap, clusterChangeHandlerCallback, helixClusterManagerCallback,
-                helixClusterManagerMetrics, dataNodeConfigSourceMetrics, sealedStateChangeCounter, partitionMap,
-                partitionNameToAmbryPartition, ambryPartitionToAmbryReplicas);
+                helixClusterManagerCallback, dataNodeConfigSourceMetrics, this);
         initializer.start();
         initializers.add(initializer);
       }
@@ -234,15 +252,7 @@ public class HelixClusterManager implements ClusterMap {
 
   @Override
   public AmbryDataNode getDataNodeId(String hostname, int port) {
-    String instanceName = getInstanceName(hostname, port);
-    AmbryDataNode dataNode = null;
-    for (DcInfo dcInfo : dcToDcInfo.values()) {
-      dataNode = dcInfo.clusterChangeHandler.getDataNode(instanceName);
-      if (dataNode != null) {
-        break;
-      }
-    }
-    return dataNode;
+    return instanceNameToAmbryDataNode.get(getInstanceName(hostname, port));
   }
 
   @Override
@@ -250,17 +260,12 @@ public class HelixClusterManager implements ClusterMap {
     if (!(dataNodeId instanceof AmbryDataNode)) {
       throw new IllegalArgumentException("Incompatible type passed in");
     }
-    AmbryDataNode dataNode = (AmbryDataNode) dataNodeId;
-    return dcToDcInfo.get(dataNode.getDatacenterName()).clusterChangeHandler.getReplicaIds(dataNode);
+    return new ArrayList<>(ambryDataNodeToAmbryReplicas.get(dataNodeId).values());
   }
 
   @Override
   public List<AmbryDataNode> getDataNodeIds() {
-    List<AmbryDataNode> dataNodeList = new ArrayList<>();
-    for (DcInfo dcInfo : dcToDcInfo.values()) {
-      dataNodeList.addAll(dcInfo.clusterChangeHandler.getAllDataNodes());
-    }
-    return dataNodeList;
+    return new ArrayList<>(instanceNameToAmbryDataNode.values());
   }
 
   @Override
@@ -314,8 +319,7 @@ public class HelixClusterManager implements ClusterMap {
       data.put(DATACENTER_NAME, dcName);
       data.put(DATACENTER_ID, dcId);
       JSONArray datanodesInDc = new JSONArray();
-      dcToDcInfo.get(dcName).clusterChangeHandler.getAllDataNodes()
-          .forEach(node -> datanodesInDc.put(node.getSnapshot()));
+      dcToNodes.get(dcName).forEach(node -> datanodesInDc.put(node.getSnapshot()));
       data.put(DATACENTER_NODES, datanodesInDc);
       datacentersJsonArray.put(data);
     });
@@ -408,14 +412,12 @@ public class HelixClusterManager implements ClusterMap {
       currentPartition = mappedPartition;
     }
     // Check if data node or disk is in current cluster map, if not, set bootstrapReplica to null.
-    ClusterChangeHandler localClusterChangeHandler =
-        dcToDcInfo.get(clusterMapConfig.clusterMapDatacenterName).clusterChangeHandler;
-    AmbryDataNode dataNode = localClusterChangeHandler.getDataNode(instanceName);
+    AmbryDataNode dataNode = instanceNameToAmbryDataNode.get(instanceName);
     String mountPathAndDiskCapacityFromHelix = replicaInfos.get(instanceName);
     String[] segments = mountPathAndDiskCapacityFromHelix.split(DISK_CAPACITY_DELIM_STR);
     String mountPath = segments[0];
     String diskCapacityStr = segments.length >= 2 ? segments[1] : null;
-    Set<AmbryDisk> disks = dataNode != null ? localClusterChangeHandler.getDisks(dataNode) : null;
+    Set<AmbryDisk> disks = dataNode != null ? ambryDataNodeToAmbryDisks.get(dataNode) : null;
     Optional<AmbryDisk> potentialDisk =
         disks != null ? disks.stream().filter(d -> d.getMountPath().equals(mountPath)).findAny() : Optional.empty();
     if (potentialDisk.isPresent()) {
@@ -475,19 +477,14 @@ public class HelixClusterManager implements ClusterMap {
   AmbryReplica getReplicaForPartitionOnNode(DataNodeId dataNodeId, String partitionString) {
     // Note: partitionString here is now from partitionId.toPathString()
     AmbryDataNode ambryDataNode = getDataNodeId(dataNodeId.getHostname(), dataNodeId.getPort());
-    return dcToDcInfo.get(dataNodeId.getDatacenterName()).clusterChangeHandler.getReplicaId(ambryDataNode,
-        partitionString);
+    return ambryDataNodeToAmbryReplicas.getOrDefault(ambryDataNode, new ConcurrentHashMap<>()).get(partitionString);
   }
 
   /**
    * @return the count of errors encountered by the Cluster Manager.
    */
   long getErrorCount() {
-    long totalErrorCnt = 0;
-    for (DcInfo dcInfo : dcToDcInfo.values()) {
-      totalErrorCnt += dcInfo.clusterChangeHandler.getErrorCount();
-    }
-    return totalErrorCnt;
+    return errorCount.get();
   }
 
   /**
@@ -502,11 +499,7 @@ public class HelixClusterManager implements ClusterMap {
    * @return a map of partition to its corresponding resource grouped by data center
    */
   Map<String, Map<String, String>> getPartitionToResourceMap() {
-    Map<String, Map<String, String>> partitionToResourceNameByDc = new HashMap<>();
-    for (DcInfo dcInfo : dcToDcInfo.values()) {
-      partitionToResourceNameByDc.put(dcInfo.dcName, dcInfo.clusterChangeHandler.getPartitionToResourceMap());
-    }
-    return partitionToResourceNameByDc;
+    return Collections.unmodifiableMap(partitionToResourceNameByDc);
   }
 
   /**
@@ -514,10 +507,6 @@ public class HelixClusterManager implements ClusterMap {
    * @return a map of data center to its data nodes
    */
   Map<String, Set<AmbryDataNode>> getDcToDataNodesMap() {
-    Map<String, Set<AmbryDataNode>> dcToNodes = new HashMap<>();
-    for (DcInfo dcInfo : dcToDcInfo.values()) {
-      dcToNodes.put(dcInfo.dcName, new HashSet<>(dcInfo.clusterChangeHandler.getAllDataNodes()));
-    }
     return Collections.unmodifiableMap(dcToNodes);
   }
 
@@ -528,10 +517,7 @@ public class HelixClusterManager implements ClusterMap {
   Map<String, RoutingTableSnapshot> getRoutingTableSnapshots() {
     Map<String, RoutingTableSnapshot> dcToRoutingTableSnapshot = new HashMap<>();
     for (DcInfo dcInfo : dcToDcInfo.values()) {
-      ClusterChangeHandler handler = dcInfo.clusterChangeHandler;
-      if (handler instanceof HelixClusterChangeHandler) {
-        dcToRoutingTableSnapshot.put(dcInfo.dcName, ((HelixClusterChangeHandler) handler).getRoutingTableSnapshot());
-      }
+      dcToRoutingTableSnapshot.put(dcInfo.dcName, dcToRoutingTableSnapshotRef.get(dcInfo.dcName).get());
     }
     return Collections.unmodifiableMap(dcToRoutingTableSnapshot);
   }
@@ -559,81 +545,74 @@ public class HelixClusterManager implements ClusterMap {
   }
 
   /**
-   * A callback class for {@link HelixClusterChangeHandler} in each dc to update cluster-wide info (i.e partition-to-replica
-   * mapping, cluster-wide capacity)
+   * Add partition if it's not present in cluster-wide partition map and also update cluster-wide allocated usable
+   * capacity. If the partition already exists, skip addition and return current partition.
+   * @param partition the {@link AmbryPartition} to add (if not present)
+   * @param capacityBytes the capacity of partition in bytes
+   * @return the current {@link AmbryPartition} present in the map.
    */
-  class ClusterChangeHandlerCallback {
-
-    /**
-     * Add partition if it's not present in cluster-wide partition map and also update cluster-wide allocated usable
-     * capacity. If the partition already exists, skip addition and return current partition.
-     * @param partition the {@link AmbryPartition} to add (if not present)
-     * @param capacityBytes the capacity of partition in bytes
-     * @return the current {@link AmbryPartition} present in the map.
-     */
-    AmbryPartition addPartitionIfAbsent(AmbryPartition partition, long capacityBytes) {
-      AmbryPartition currentPartition = partitionNameToAmbryPartition.putIfAbsent(partition.toPathString(), partition);
-      if (currentPartition == null) {
-        // this means the map previously didn't contain this partition and passed-in partition is successfully added
-        // into the map
-        currentPartition = partition;
-        // it doesn't really need to synchronize this method. "partitionNameToAmbryPartition" guarantees each thread
-        // will get same instance of ambry partition. The first one that succeeds adding partition into
-        // "partitionNameToAmbryPartition" will update "partitionMap".
-        partitionMap.put(ByteBuffer.wrap(currentPartition.getBytes()), currentPartition);
-        // update cluster-wide capacity
-        clusterWideAllocatedUsableCapacityBytes.getAndAdd(capacityBytes);
-      }
-      ambryPartitionToAmbryReplicas.putIfAbsent(currentPartition, ConcurrentHashMap.newKeySet());
-      return currentPartition;
+  private AmbryPartition addPartitionIfAbsent(AmbryPartition partition, long capacityBytes) {
+    AmbryPartition currentPartition = partitionNameToAmbryPartition.putIfAbsent(partition.toPathString(), partition);
+    if (currentPartition == null) {
+      // this means the map previously didn't contain this partition and passed-in partition is successfully added
+      // into the map
+      currentPartition = partition;
+      // it doesn't really need to synchronize this method. "partitionNameToAmbryPartition" guarantees each thread
+      // will get same instance of ambry partition. The first one that succeeds adding partition into
+      // "partitionNameToAmbryPartition" will update "partitionMap".
+      partitionMap.put(ByteBuffer.wrap(currentPartition.getBytes()), currentPartition);
+      // update cluster-wide capacity
+      clusterWideAllocatedUsableCapacityBytes.getAndAdd(capacityBytes);
     }
+    ambryPartitionToAmbryReplicas.putIfAbsent(currentPartition, ConcurrentHashMap.newKeySet());
+    return currentPartition;
+  }
 
-    /**
-     * Add a list of {@link AmbryReplica} to given {@link AmbryPartition} and update cluster-wide capacity stats
-     * @param partition the {@link AmbryPartition} which replicas should be added to.
-     * @param replicas list of {@link AmbryReplica} to be added.
-     */
-    void addReplicasToPartition(AmbryPartition partition, List<AmbryReplica> replicas) {
-      AmbryPartition currentPartition = addPartitionIfAbsent(partition, replicas.get(0).getCapacityInBytes());
-      ambryPartitionToAmbryReplicas.compute(currentPartition, (k, v) -> {
-        // calling addPartitionIfAbsent guarantees that v is not null
-        v.addAll(replicas);
-        return v;
-      });
-      clusterWideAllocatedRawCapacityBytes.getAndAdd(replicas.get(0).getCapacityInBytes() * replicas.size());
-    }
+  /**
+   * Add a list of {@link AmbryReplica} to given {@link AmbryPartition} and update cluster-wide capacity stats
+   * @param partition the {@link AmbryPartition} which replicas should be added to.
+   * @param replicas list of {@link AmbryReplica} to be added.
+   */
+  private void addReplicasToPartition(AmbryPartition partition, List<AmbryReplica> replicas) {
+    AmbryPartition currentPartition = addPartitionIfAbsent(partition, replicas.get(0).getCapacityInBytes());
+    ambryPartitionToAmbryReplicas.compute(currentPartition, (k, v) -> {
+      // calling addPartitionIfAbsent guarantees that v is not null
+      v.addAll(replicas);
+      return v;
+    });
+    clusterWideAllocatedRawCapacityBytes.getAndAdd(replicas.get(0).getCapacityInBytes() * replicas.size());
+  }
 
-    /**
-     * Remove {@link AmbryReplica}(s) from given {@link AmbryPartition} and update cluster-wide capacity stats
-     * @param partition the {@link AmbryPartition} which replicas should be removed from.
-     * @param replicas list of {@link AmbryReplica} to be removed.
-     */
-    void removeReplicasFromPartition(AmbryPartition partition, List<AmbryReplica> replicas) {
-      ambryPartitionToAmbryReplicas.computeIfPresent(partition, (k, v) -> {
-        v.removeAll(replicas);
-        clusterWideAllocatedRawCapacityBytes.getAndAdd(-1 * replicas.get(0).getCapacityInBytes() * replicas.size());
-        return v;
-      });
-    }
+  /**
+   * Remove {@link AmbryReplica}(s) from given {@link AmbryPartition} and update cluster-wide capacity stats
+   * @param partition the {@link AmbryPartition} which replicas should be removed from.
+   * @param replicas list of {@link AmbryReplica} to be removed.
+   */
+  private void removeReplicasFromPartition(AmbryPartition partition, List<AmbryReplica> replicas) {
+    ambryPartitionToAmbryReplicas.computeIfPresent(partition, (k, v) -> {
+      v.removeAll(replicas);
+      clusterWideAllocatedRawCapacityBytes.getAndAdd(-1 * replicas.get(0).getCapacityInBytes() * replicas.size());
+      return v;
+    });
+  }
 
-    /**
-     * Add cluster-wide raw capacity. This is called when new disk is created and added to cluster. We update
-     * cluster-wide raw capacity by adding the disk capacity.
-     * @param diskRawCapacityBytes raw disk capacity to be added.
-     */
-    void addClusterWideRawCapacity(long diskRawCapacityBytes) {
-      clusterWideRawCapacityBytes.getAndAdd(diskRawCapacityBytes);
-    }
+  /**
+   * Add cluster-wide raw capacity. This is called when new disk is created and added to cluster. We update
+   * cluster-wide raw capacity by adding the disk capacity.
+   * @param diskRawCapacityBytes raw disk capacity to be added.
+   */
+  private void addClusterWideRawCapacity(long diskRawCapacityBytes) {
+    clusterWideRawCapacityBytes.getAndAdd(diskRawCapacityBytes);
+  }
 
-    /**
-     * Pop out bootstrap replica (if any) on current instance. A bootstrap replica is a replica dynamically added to
-     * current node at runtime.
-     * @param partitionName the partition name of bootstrap replica.
-     * @return bootstrap replica or {@code null} if not found.
-     */
-    AmbryReplica fetchBootstrapReplica(String partitionName) {
-      return (AmbryReplica) bootstrapReplicas.remove(partitionName);
-    }
+  /**
+   * Pop out bootstrap replica (if any) on current instance. A bootstrap replica is a replica dynamically added to
+   * current node at runtime.
+   * @param partitionName the partition name of bootstrap replica.
+   * @return bootstrap replica or {@code null} if not found.
+   */
+  private AmbryReplica fetchBootstrapReplica(String partitionName) {
+    return (AmbryReplica) bootstrapReplicas.remove(partitionName);
   }
 
   /**
@@ -655,8 +634,8 @@ public class HelixClusterManager implements ClusterMap {
     public String getResourceNameForPartition(AmbryPartition partition) {
       String result = null;
       // go through all datacenters in case that partition resides in one datacenter only
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        String resourceName = dcInfo.clusterChangeHandler.getPartitionToResourceMap().get(partition.toPathString());
+      for (String dcName : partitionToResourceNameByDc.keySet()) {
+        String resourceName = partitionToResourceNameByDc.get(dcName).get(partition.toPathString());
         if (resourceName != null) {
           result = resourceName;
           break;
@@ -676,7 +655,15 @@ public class HelixClusterManager implements ClusterMap {
       for (DcInfo dcInfo : dcToDcInfo.values()) {
         String dc = dcInfo.dcName;
         if (dcName == null || dcName.equals(dc)) {
-          dcInfo.clusterChangeHandler.getReplicaIdsByState(partition, state).forEach(replicas::add);
+          String resourceName = partitionToResourceNameByDc.get(dc).get(partition.toPathString());
+          dcToRoutingTableSnapshotRef.get(dcInfo.dcName)
+              .get()
+              .getInstancesForResource(resourceName, partition.toPathString(), state.name())
+              .stream()
+              .map(instanceConfig -> instanceNameToAmbryDataNode.get(instanceConfig.getInstanceName()))
+              .map(dataNode -> ambryDataNodeToAmbryReplicas.get(dataNode).get(partition.toPathString()))
+              .filter(Objects::nonNull)
+              .forEach(replicas::add);
         }
       }
       return replicas;
@@ -688,7 +675,17 @@ public class HelixClusterManager implements ClusterMap {
       for (DcInfo dcInfo : dcToDcInfo.values()) {
         String dc = dcInfo.dcName;
         if (dcName == null || dcName.equals(dc)) {
-          dcInfo.clusterChangeHandler.getReplicaIdsByStates(replicasByState, states, partition);
+          String resourceName = partitionToResourceNameByDc.get(dc).get(partition.toPathString());
+          RoutingTableSnapshot snapshot = dcToRoutingTableSnapshotRef.get(dc).get();
+          for (ReplicaState state : states) {
+            List<AmbryReplica> list = replicasByState.computeIfAbsent(state, k -> new ArrayList<>());
+            snapshot.getInstancesForResource(resourceName, partition.toPathString(), state.name())
+                .stream()
+                .map(instanceConfig -> instanceNameToAmbryDataNode.get(instanceConfig.getInstanceName()))
+                .map(dataNode -> ambryDataNodeToAmbryReplicas.get(dataNode).get(partition.toPathString()))
+                .filter(Objects::nonNull)
+                .forEach(list::add);
+          }
         }
       }
     }
@@ -739,13 +736,11 @@ public class HelixClusterManager implements ClusterMap {
     @Override
     public Collection<AmbryDisk> getDisks(AmbryDataNode dataNode) {
       if (dataNode != null) {
-        return dcToDcInfo.get(dataNode.getDatacenterName()).clusterChangeHandler.getDisks(dataNode);
+        return ambryDataNodeToAmbryDisks.get(dataNode);
       }
       List<AmbryDisk> disksToReturn = new ArrayList<>();
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        for (Set<AmbryDisk> disks : dcInfo.clusterChangeHandler.getDataNodeToDisksMap().values()) {
-          disksToReturn.addAll(disks);
-        }
+      for (Set<AmbryDisk> disks : ambryDataNodeToAmbryDisks.values()) {
+        disksToReturn.addAll(disks);
       }
       return disksToReturn;
     }
@@ -755,10 +750,8 @@ public class HelixClusterManager implements ClusterMap {
      */
     long getDiskCount() {
       long count = 0;
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        for (Set<AmbryDisk> disks : dcInfo.clusterChangeHandler.getDataNodeToDisksMap().values()) {
-          count += disks.size();
-        }
+      for (Set<AmbryDisk> disks : ambryDataNodeToAmbryDisks.values()) {
+        count += disks.size();
       }
       return count;
     }
@@ -836,6 +829,462 @@ public class HelixClusterManager implements ClusterMap {
      */
     long getAllocatedUsableCapacity() {
       return clusterWideAllocatedUsableCapacityBytes.get();
+    }
+
+    /**
+     * Gets the {@link RoutingTableSnapshot} for a given data center or entire cluster.
+     * @param dcName data center name.
+     * @return {@link RoutingTableSnapshot} for a given data center. If the input data center name is null or if using
+     * aggregated view, returns the routing table snapshot of entire cluster.
+     */
+    public RoutingTableSnapshot getRoutingTableSnapshot(String dcName) {
+      return dcToRoutingTableSnapshotRef.get(dcName).get();
+    }
+  }
+
+  /**
+   * A more dynamic implementation of {@link HelixClusterChangeHandler} which supports adding new nodes/partitions at
+   * runtime. It is also able to absorb replica location changes in cluster.
+   */
+  class HelixClusterChangeHandler
+      implements DataNodeConfigChangeListener, LiveInstanceChangeListener, IdealStateChangeListener,
+                 RoutingTableChangeListener {
+    private final String dcName;
+    private final Object notificationLock = new Object();
+    private final Consumer<Exception> onInitializationFailure;
+    private final HelixClusterManager.HelixClusterManagerCallback helixClusterManagerCallback;
+    private final CountDownLatch routingTableInitLatch = new CountDownLatch(1);
+    private final List<ClusterMapChangeListener> clusterMapChangeListeners = new ArrayList<>();
+    private volatile boolean instanceConfigInitialized = false;
+    private volatile boolean liveStateInitialized = false;
+    private volatile boolean idealStateInitialized = false;
+
+    final Set<String> allInstances = new HashSet<>();
+
+    /**
+     * @param dcName the name of data center this handler is associated with.
+     * @param helixClusterManagerCallback a call back used to query cluster-wide info.
+     * @param onInitializationFailure callback to be called if initialization fails in a listener call.
+     */
+    HelixClusterChangeHandler(String dcName,
+        HelixClusterManager.HelixClusterManagerCallback helixClusterManagerCallback,
+        Consumer<Exception> onInitializationFailure) {
+      this.dcName = dcName;
+      this.helixClusterManagerCallback = helixClusterManagerCallback;
+      this.onInitializationFailure = onInitializationFailure;
+    }
+
+    /**
+     * Handle any {@link DataNodeConfig} related change in current datacenter. Several events will trigger instance config
+     * change: (1) replica's seal or stop state has changed; (2) new node or new partition is added; (3) new replica is
+     * added to existing node; (4) old replica is removed from existing node; (5) data node is deleted from cluster.
+     * For now, we support (1)~(4). We may consider supporting (5) in the future.
+     * (The ZNode path of instance config in Helix is [AmbryClusterName]/CONFIGS/PARTICIPANT/[hostname_port])
+     * @param configs all the {@link DataNodeConfig}(s) in current data center. (Note that PreFetch is enabled by default
+     *                in Helix, which means all instance configs under "participants" ZNode will be sent to this method)
+     */
+    @Override
+    public void onDataNodeConfigChange(Iterable<DataNodeConfig> configs) {
+      try {
+        synchronized (notificationLock) {
+          if (!instanceConfigInitialized) {
+            logger.info("Received initial notification for instance config change from {}", dcName);
+          } else {
+            logger.info("Instance config change triggered from {}", dcName);
+          }
+          if (logger.isDebugEnabled()) {
+            configs.forEach(config -> logger.debug("Detailed data node config in {} is: {}", dcName, config));
+          }
+          try {
+            addOrUpdateInstanceInfos(configs);
+          } catch (Exception e) {
+            if (!instanceConfigInitialized) {
+              logger.error("Exception occurred when initializing instances in {}: ", dcName, e);
+              onInitializationFailure.accept(e);
+            } else {
+              logger.error("Exception occurred at runtime when handling instance config changes in {}: ", dcName, e);
+              helixClusterManagerMetrics.instanceConfigChangeErrorCount.inc();
+            }
+          } finally {
+            instanceConfigInitialized = true;
+          }
+          sealedStateChangeCounter.incrementAndGet();
+          helixClusterManagerMetrics.instanceConfigChangeTriggerCount.inc();
+        }
+      } catch (Throwable t) {
+        errorCount.incrementAndGet();
+        throw t;
+      }
+    }
+
+    /**
+     * Triggered whenever the IdealState in current data center has changed (for now, it is usually updated by Helix
+     * Bootstrap tool).
+     * @param idealState a list of {@link IdealState} that specifies ideal location of replicas.
+     * @param changeContext the {@link NotificationContext} associated.
+     */
+    @Override
+    public void onIdealStateChange(List<IdealState> idealState, NotificationContext changeContext) {
+      if (!idealStateInitialized) {
+        logger.info("Received initial notification for IdealState change from {}", dcName);
+        idealStateInitialized = true;
+      } else {
+        logger.info("IdealState change triggered from {}", dcName);
+      }
+      logger.debug("Detailed ideal states in {} are: {}", dcName, idealState);
+      // rebuild the entire partition-to-resource map in current dc
+      ConcurrentHashMap<String, String> partitionToResourceMap = new ConcurrentHashMap<>();
+      for (IdealState state : idealState) {
+        String resourceName = state.getResourceName();
+        state.getPartitionSet().forEach(partitionName -> partitionToResourceMap.put(partitionName, resourceName));
+      }
+      partitionToResourceNameByDc.put(dcName, partitionToResourceMap);
+      helixClusterManagerMetrics.idealStateChangeTriggerCount.inc();
+    }
+
+    /**
+     * Triggered whenever there is a change in the list of live instances.
+     * @param liveInstances the list of all live instances (not a change set) at the time of this call.
+     * @param changeContext the {@link NotificationContext} associated.
+     */
+    @Override
+    public void onLiveInstanceChange(List<LiveInstance> liveInstances, NotificationContext changeContext) {
+      try {
+        if (!liveStateInitialized) {
+          logger.info("Received initial notification for live instance change from {}", dcName);
+          liveStateInitialized = true;
+        } else {
+          logger.info("Live instance change triggered from {}", dcName);
+        }
+        logger.debug("Detailed live instances in {} are: {}", dcName, liveInstances);
+        synchronized (notificationLock) {
+          updateInstanceLiveness(liveInstances);
+          helixClusterManagerMetrics.liveInstanceChangeTriggerCount.inc();
+        }
+      } catch (Throwable t) {
+        errorCount.incrementAndGet();
+        throw t;
+      }
+    }
+
+    /**
+     * Triggered whenever the state of replica in cluster has changed. The snapshot contains up-to-date state of all
+     * resources(replicas) in this data center.
+     * @param routingTableSnapshot a snapshot of routing table for this data center.
+     * @param context additional context associated with this change.
+     */
+    @Override
+    public void onRoutingTableChange(RoutingTableSnapshot routingTableSnapshot, Object context) {
+      dcToRoutingTableSnapshotRef.computeIfAbsent(dcName, k -> new AtomicReference<>()).getAndSet(routingTableSnapshot);
+      if (routingTableInitLatch.getCount() == 1) {
+        logger.info("Received initial notification for routing table change from {}", dcName);
+        routingTableInitLatch.countDown();
+      } else {
+        logger.info("Routing table change triggered from {}", dcName);
+      }
+
+      //we should notify routing table change indication to different cluster map change listeners (i.e replication manager, partition selection helper, etc). However, only replication manager handles this callback now.
+      //On receiving this notification, listeners can query the latest state information of replicas by querying APIs provided in class RoutingTableSnapshot
+      for (ClusterMapChangeListener listener : clusterMapChangeListeners) {
+        listener.onRoutingTableChange();
+      }
+
+      helixClusterManagerMetrics.routingTableChangeTriggerCount.inc();
+    }
+
+    /**
+     * Sets the helix {@link RoutingTableSnapshot} for entire cluster or a given data center.
+     * @param routingTableSnapshot snapshot of cluster mappings.
+     */
+    public void setRoutingTableSnapshot(RoutingTableSnapshot routingTableSnapshot) {
+      dcToRoutingTableSnapshotRef.computeIfAbsent(dcName, k -> new AtomicReference<>()).getAndSet(routingTableSnapshot);
+    }
+
+    public void waitForInitNotification() throws InterruptedException {
+      // wait slightly more than 5 mins to ensure routerUpdater refreshes the snapshot.
+      if (!routingTableInitLatch.await(320, TimeUnit.SECONDS)) {
+        throw new IllegalStateException("Initial routing table change from " + dcName + " didn't come within 5 mins");
+      }
+    }
+
+    public void registerClusterMapListener(ClusterMapChangeListener clusterMapChangeListener) {
+      clusterMapChangeListeners.add(clusterMapChangeListener);
+    }
+
+    /**
+     * Add new instances or update existing instances based on {@link DataNodeConfig}(s). This may also invoke callbacks
+     * in some clustermap change listeners (i.e. {@link PartitionSelectionHelper}, ReplicationManager)
+     * @param dataNodeConfigs the {@link DataNodeConfig}(s) used to update in-mem cluster map.
+     */
+    private void addOrUpdateInstanceInfos(Iterable<DataNodeConfig> dataNodeConfigs) throws Exception {
+      List<ReplicaId> totalAddedReplicas = new ArrayList<>();
+      List<ReplicaId> totalRemovedReplicas = new ArrayList<>();
+      for (DataNodeConfig dataNodeConfig : dataNodeConfigs) {
+        Pair<List<ReplicaId>, List<ReplicaId>> addedAndRemovedReplicas;
+        if (instanceNameToAmbryDataNode.containsKey(dataNodeConfig.getInstanceName())) {
+          addedAndRemovedReplicas = updateInstanceInfo(dataNodeConfig);
+        } else {
+          addedAndRemovedReplicas = new Pair<>(createNewInstance(dataNodeConfig), new ArrayList<>());
+        }
+        totalAddedReplicas.addAll(addedAndRemovedReplicas.getFirst());
+        totalRemovedReplicas.addAll(addedAndRemovedReplicas.getSecond());
+      }
+      // if this is not initial InstanceConfig change and any replicas are added or removed, we should invoke callbacks
+      // for different clustermap change listeners (i.e replication manager, partition selection helper)
+      logger.info(
+          "In total, {} replicas are being added and {} replicas are being removed. instanceConfigInitialized: {}",
+          totalAddedReplicas.size(), totalRemovedReplicas.size(), instanceConfigInitialized);
+      if (instanceConfigInitialized && (!totalAddedReplicas.isEmpty() || !totalRemovedReplicas.isEmpty())) {
+        for (ClusterMapChangeListener listener : clusterMapChangeListeners) {
+          listener.onReplicaAddedOrRemoved(totalAddedReplicas, totalRemovedReplicas);
+        }
+      }
+    }
+
+    /**
+     * Update info of an existing instance. This may happen in following cases: (1) new replica is added; (2) old replica
+     * is removed; (3) replica's state has changed (i.e. becomes seal/unseal).
+     * @param dataNodeConfig the {@link DataNodeConfig} used to update info of instance.
+     * @return a pair of lists: (1) new added replicas; (2) removed old replicas, during this update.
+     */
+    private Pair<List<ReplicaId>, List<ReplicaId>> updateInstanceInfo(DataNodeConfig dataNodeConfig) throws Exception {
+      final List<ReplicaId> addedReplicas = new ArrayList<>();
+      final List<ReplicaId> removedReplicas = new ArrayList<>();
+      String instanceName = dataNodeConfig.getInstanceName();
+      logger.info("Updating replicas info for existing node {}", instanceName);
+      Set<String> sealedReplicas = dataNodeConfig.getSealedReplicas();
+      Set<String> stoppedReplicas = dataNodeConfig.getStoppedReplicas();
+      AmbryDataNode dataNode = instanceNameToAmbryDataNode.get(instanceName);
+      ConcurrentHashMap<String, AmbryReplica> currentReplicasOnNode = ambryDataNodeToAmbryReplicas.get(dataNode);
+      ConcurrentHashMap<String, AmbryReplica> replicasFromInstanceConfig = new ConcurrentHashMap<>();
+      Map<String, AmbryDisk> mountPathToDisk = ambryDataNodeToAmbryDisks.get(dataNode)
+          .stream()
+          .collect(Collectors.toMap(AmbryDisk::getMountPath, disk -> disk));
+      Map<AmbryPartition, List<AmbryReplica>> replicaToAddByPartition = new HashMap<>();
+      for (Map.Entry<String, DataNodeConfig.DiskConfig> diskEntry : dataNodeConfig.getDiskConfigs().entrySet()) {
+        String mountPath = diskEntry.getKey();
+        DataNodeConfig.DiskConfig diskConfig = diskEntry.getValue();
+        AmbryDisk disk = mountPathToDisk.getOrDefault(mountPath, null);
+        if (disk == null) {
+          logger.warn("{} is a new disk or unrecognizable disk which is not supported on existing node {}.", mountPath,
+              instanceName);
+          // TODO support dynamically adding disk in the future
+          continue;
+        }
+        // update disk capacity if needed
+        if (disk.getRawCapacityInBytes() != diskConfig.getDiskCapacityInBytes()) {
+          long prevDiskCapacity = disk.getRawCapacityInBytes();
+          logger.info("Capacity of disk at {} on {} has changed. Previous was: {} bytes, new capacity is {} bytes",
+              mountPath, instanceName, prevDiskCapacity, diskConfig.getDiskCapacityInBytes());
+          disk.setDiskCapacityInBytes(diskConfig.getDiskCapacityInBytes());
+          addClusterWideRawCapacity(diskConfig.getDiskCapacityInBytes() - prevDiskCapacity);
+        }
+        for (Map.Entry<String, DataNodeConfig.ReplicaConfig> replicaEntry : diskConfig.getReplicaConfigs().entrySet()) {
+          // partition name and replica name are the same.
+          String partitionName = replicaEntry.getKey();
+          DataNodeConfig.ReplicaConfig replicaConfig = replicaEntry.getValue();
+          if (currentReplicasOnNode.containsKey(partitionName)) {
+            // if replica is already present
+            AmbryReplica existingReplica = currentReplicasOnNode.get(partitionName);
+            // 1. directly add it into "replicasFromInstanceConfig" map
+            replicasFromInstanceConfig.put(partitionName, existingReplica);
+            // 2. update replica seal/stop state
+            updateReplicaStateAndOverrideIfNeeded(existingReplica, sealedReplicas, stoppedReplicas);
+          } else {
+            // if this is a new replica and doesn't exist on node
+            logger.info("Adding new replica {} to existing node {} in {}", partitionName, instanceName, dcName);
+            // this can be a brand new partition that is added to an existing node
+            AmbryPartition mappedPartition =
+                new AmbryPartition(Long.parseLong(partitionName), replicaConfig.getPartitionClass(),
+                    helixClusterManagerCallback);
+            // Ensure only one AmbryPartition instance exists for specific partition.
+            mappedPartition = addPartitionIfAbsent(mappedPartition, replicaConfig.getReplicaCapacityInBytes());
+            ensurePartitionAbsenceOnNodeAndValidateCapacity(mappedPartition, dataNode,
+                replicaConfig.getReplicaCapacityInBytes());
+            // create new replica belonging to this partition or find the existing replica from bootstrapReplicas map.
+            AmbryReplica replica;
+            if (selfInstanceName.equals(instanceName)) {
+              // if this is a newly added replica on current instance, it should be present in bootstrapReplicas map.
+              replica = fetchBootstrapReplica(mappedPartition.toPathString());
+              if (replica == null) {
+                logger.error("Replica {} is not present in bootstrap replica set, abort instance info update",
+                    mappedPartition.toPathString());
+                throw new IllegalStateException("Replica to add is not present in bootstrap replica map");
+              }
+            } else {
+              replica = new AmbryServerReplica(clusterMapConfig, mappedPartition, disk,
+                  stoppedReplicas.contains(partitionName), replicaConfig.getReplicaCapacityInBytes(),
+                  sealedReplicas.contains(partitionName));
+            }
+            updateReplicaStateAndOverrideIfNeeded(replica, sealedReplicas, stoppedReplicas);
+            // add new created replica to "replicasFromInstanceConfig" map
+            replicasFromInstanceConfig.put(partitionName, replica);
+            // Put new replica into partition-to-replica map temporarily (this is to avoid any exception thrown within the
+            // loop before updating "ambryDataNodeToAmbryReplicas" map. If we update call addReplicasToPartition here
+            // immediately, the exception may cause inconsistency between "ambryPartitionToAmbryReplicas" and
+            // "ambryDataNodeToAmbryReplicas")
+            replicaToAddByPartition.put(mappedPartition, Collections.singletonList(replica));
+          }
+        }
+      }
+      // update "ambryDataNodeToAmbryReplicas" map and "addedReplicas" list
+      replicaToAddByPartition.forEach((k, v) -> {
+        addReplicasToPartition(k, v);
+        addedReplicas.addAll(v);
+      });
+      // update ambryDataNodeToAmbryReplicas map by adding "replicasFromInstanceConfig"
+      ambryDataNodeToAmbryReplicas.put(instanceNameToAmbryDataNode.get(instanceName), replicasFromInstanceConfig);
+      // Derive old replicas that are removed and delete them from partition
+      currentReplicasOnNode.keySet()
+          .stream()
+          .filter(partitionName -> !replicasFromInstanceConfig.containsKey(partitionName))
+          .forEach(pName -> {
+            logger.info("Removing replica {} from existing node {}", pName, instanceName);
+            AmbryReplica ambryReplica = currentReplicasOnNode.get(pName);
+            removeReplicasFromPartition(ambryReplica.getPartitionId(), Collections.singletonList(ambryReplica));
+            removedReplicas.add(ambryReplica);
+          });
+      return new Pair<>(addedReplicas, removedReplicas);
+    }
+
+    /**
+     * If partition override is enabled, we override replica SEAL/UNSEAL state based on partitionOverrideMap. If disabled,
+     * update replica state according to the info from {@link DataNodeConfig}.
+     * @param replica the {@link ReplicaId} whose states (seal,stop) should be updated.
+     * @param sealedReplicas a collection of {@link ReplicaId}(s) that are in SEALED state.
+     * @param stoppedReplicas a collection of {@link ReplicaId}(s) that are in STOPPED state.
+     */
+    private void updateReplicaStateAndOverrideIfNeeded(AmbryReplica replica, Collection<String> sealedReplicas,
+        Collection<String> stoppedReplicas) {
+      String partitionName = replica.getPartitionId().toPathString();
+      boolean isSealed;
+      if (clusterMapConfig.clusterMapEnablePartitionOverride && partitionOverrideInfoMap.containsKey(partitionName)) {
+        isSealed = partitionOverrideInfoMap.get(partitionName)
+            .get(ClusterMapUtils.PARTITION_STATE)
+            .equals(ClusterMapUtils.READ_ONLY_STR);
+      } else {
+        isSealed = sealedReplicas.contains(partitionName);
+      }
+      replica.setSealedState(isSealed);
+      replica.setStoppedState(stoppedReplicas.contains(partitionName));
+    }
+
+    /**
+     * Create a new instance(node) and initialize disks/replicas on it.
+     * @param dataNodeConfig the {@link DataNodeConfig} to create new instance
+     * @return a list of newly added replicas;
+     * @throws Exception if there is an exception in instantiating the {@link ResourceStatePolicy}
+     */
+    private List<ReplicaId> createNewInstance(DataNodeConfig dataNodeConfig) throws Exception {
+      String instanceName = dataNodeConfig.getInstanceName();
+      logger.info("Adding node {} and its disks and replicas in {}", instanceName, dcName);
+      AmbryDataNode datanode =
+          new AmbryServerDataNode(dataNodeConfig.getDatacenterName(), clusterMapConfig, dataNodeConfig.getHostName(),
+              dataNodeConfig.getPort(), dataNodeConfig.getRackId(), dataNodeConfig.getSslPort(),
+              dataNodeConfig.getHttp2Port(), DEFAULT_XID, helixClusterManagerCallback);
+      // for new instance, we first set it to unavailable and rely on its participation to update its liveness
+      if (!instanceName.equals(selfInstanceName)) {
+        datanode.setState(HardwareState.UNAVAILABLE);
+      }
+      List<ReplicaId> addedReplicas = initializeDisksAndReplicasOnNode(datanode, dataNodeConfig);
+      instanceNameToAmbryDataNode.put(instanceName, datanode);
+      dcToNodes.computeIfAbsent(datanode.getDatacenterName(), s -> new HashSet<>()).add(datanode);
+      allInstances.add(instanceName);
+      return addedReplicas;
+    }
+
+    /**
+     * Initialize the disks and replicas on the given node. Create partitions if this is the first time a replica of
+     * that partition is being constructed. If partition override is enabled, the seal state of replica is determined by
+     * partition info in HelixPropertyStore, if disabled, the seal state is determined by {@code dataNodeConfig}.
+     * @param datanode the {@link AmbryDataNode} that is being initialized.
+     * @param dataNodeConfig the {@link DataNodeConfig} associated with this datanode.
+     * @return a list of newly added replicas on this node.
+     * @throws Exception if creation of {@link AmbryDisk} throws an Exception.
+     */
+    private List<ReplicaId> initializeDisksAndReplicasOnNode(AmbryDataNode datanode, DataNodeConfig dataNodeConfig)
+        throws Exception {
+      List<ReplicaId> addedReplicas = new ArrayList<>();
+      Set<String> sealedReplicas = dataNodeConfig.getSealedReplicas();
+      Set<String> stoppedReplicas = dataNodeConfig.getStoppedReplicas();
+      ambryDataNodeToAmbryReplicas.put(datanode, new ConcurrentHashMap<>());
+      ambryDataNodeToAmbryDisks.put(datanode, new HashSet<>());
+      for (Map.Entry<String, DataNodeConfig.DiskConfig> diskEntry : dataNodeConfig.getDiskConfigs().entrySet()) {
+        String mountPath = diskEntry.getKey();
+        DataNodeConfig.DiskConfig diskConfig = diskEntry.getValue();
+
+        // Create disk
+        AmbryDisk disk = new AmbryDisk(clusterMapConfig, datanode, mountPath, diskConfig.getState(),
+            diskConfig.getDiskCapacityInBytes());
+        ambryDataNodeToAmbryDisks.get(datanode).add(disk);
+        addClusterWideRawCapacity(diskConfig.getDiskCapacityInBytes());
+
+        for (Map.Entry<String, DataNodeConfig.ReplicaConfig> replicaEntry : diskConfig.getReplicaConfigs().entrySet()) {
+          String partitionName = replicaEntry.getKey();
+          DataNodeConfig.ReplicaConfig replicaConfig = replicaEntry.getValue();
+
+          AmbryPartition mappedPartition =
+              new AmbryPartition(Long.parseLong(partitionName), replicaConfig.getPartitionClass(),
+                  helixClusterManagerCallback);
+          // Ensure only one AmbryPartition instance exists for specific partition.
+          mappedPartition = addPartitionIfAbsent(mappedPartition, replicaConfig.getReplicaCapacityInBytes());
+          ensurePartitionAbsenceOnNodeAndValidateCapacity(mappedPartition, datanode,
+              replicaConfig.getReplicaCapacityInBytes());
+          // Create replica associated with this node and this partition
+          boolean isSealed = sealedReplicas.contains(partitionName);
+          if (clusterMapConfig.clusterMapEnablePartitionOverride && partitionOverrideInfoMap.containsKey(
+              partitionName)) {
+            // override sealed state if PartitionOverride is enabled.
+            isSealed = partitionOverrideInfoMap.get(partitionName)
+                .get(ClusterMapUtils.PARTITION_STATE)
+                .equals(ClusterMapUtils.READ_ONLY_STR);
+          }
+          AmbryReplica replica =
+              new AmbryServerReplica(clusterMapConfig, mappedPartition, disk, stoppedReplicas.contains(partitionName),
+                  replicaConfig.getReplicaCapacityInBytes(), isSealed);
+          ambryDataNodeToAmbryReplicas.get(datanode).put(mappedPartition.toPathString(), replica);
+          addReplicasToPartition(mappedPartition, Collections.singletonList(replica));
+          addedReplicas.add(replica);
+        }
+      }
+      return addedReplicas;
+    }
+
+    /**
+     * Update the liveness states of existing instances based on the input.
+     * @param liveInstances the list of instances that are up.
+     */
+    private void updateInstanceLiveness(List<LiveInstance> liveInstances) {
+      Set<String> liveInstancesSet = new HashSet<>();
+      liveInstances.forEach(e -> liveInstancesSet.add(e.getInstanceName()));
+      for (String instanceName : allInstances) {
+        // Here we ignore live instance change it's about self instance. The reason is, during server's startup, current
+        // node should be AVAILABLE but the list of live instances doesn't include current node since it hasn't joined yet.
+        if (liveInstancesSet.contains(instanceName) || instanceName.equals(selfInstanceName)) {
+          instanceNameToAmbryDataNode.get(instanceName).setState(HardwareState.AVAILABLE);
+        } else {
+          instanceNameToAmbryDataNode.get(instanceName).setState(HardwareState.UNAVAILABLE);
+        }
+      }
+    }
+
+    /**
+     * Ensure that the given partition is absent on the given datanode. This is called as part of an inline validation
+     * done to ensure that two replicas of the same partition do not exist on the same datanode.
+     * @param partition the {@link AmbryPartition} to check.
+     * @param datanode the {@link AmbryDataNode} on which to check.
+     * @param expectedReplicaCapacity the capacity expected for the replicas of the partition.
+     */
+    private void ensurePartitionAbsenceOnNodeAndValidateCapacity(AmbryPartition partition, AmbryDataNode datanode,
+        long expectedReplicaCapacity) {
+      for (AmbryReplica replica : helixClusterManagerCallback.getReplicaIdsForPartition(partition)) {
+        if (replica.getDataNodeId().equals(datanode)) {
+          throw new IllegalStateException("Replica already exists on " + datanode + " for " + partition);
+        } else if (replica.getCapacityInBytes() != expectedReplicaCapacity) {
+          throw new IllegalStateException("Expected replica capacity " + expectedReplicaCapacity + " is different from "
+              + "the capacity of an existing replica " + replica.getCapacityInBytes());
+        }
+      }
     }
   }
 }
