+++ ambry-clustermap/src/test/java/com/github/ambry/clustermap/ClusterChangeHandlerTest.java
@@ -15,6 +15,7 @@ package com.github.ambry.clustermap;
 
 import com.codahale.metrics.Counter;
 import com.codahale.metrics.MetricRegistry;
+import com.github.ambry.clustermap.HelixClusterManager.HelixClusterChangeHandler;
 import com.github.ambry.config.ClusterMapConfig;
 import com.github.ambry.config.VerifiableProperties;
 import com.github.ambry.utils.Utils;
@@ -53,7 +54,7 @@ import static org.mockito.Mockito.*;
 
 
 /**
- * Test {@link DynamicClusterChangeHandler} to verify its behavior is expected when handling cluster changes.
+ * Test {@link HelixClusterChangeHandler} to verify its behavior is expected when handling cluster changes.
  */
 @RunWith(Parameterized.class)
 public class ClusterChangeHandlerTest {
@@ -213,26 +214,21 @@ public class ClusterChangeHandlerTest {
   }
 
   /**
-   * Test that {@link DynamicClusterChangeHandler} is able to handle invalid info entry in the InstanceConfig at runtime
+   * Test that {@link HelixClusterChangeHandler} is able to handle invalid info entry in the InstanceConfig at runtime
    * or during initialization.
    */
   @Test
-  public void instanceConfigInvalidInfoEntryTest() {
+  public void instanceConfigInvalidInfoEntryTest() throws IOException {
     Properties properties = new Properties();
     properties.putAll(props);
     ClusterMapConfig clusterMapConfig = new ClusterMapConfig(new VerifiableProperties(properties));
-    HelixClusterManager.HelixClusterManagerCallback mockManagerCallback =
-        Mockito.mock(HelixClusterManager.HelixClusterManagerCallback.class);
-    HelixClusterManager.ClusterChangeHandlerCallback mockHandlerCallback =
-        Mockito.mock(HelixClusterManager.ClusterChangeHandlerCallback.class);
-    doAnswer(returnsFirstArg()).when(mockHandlerCallback).addPartitionIfAbsent(any(), anyLong());
+    HelixClusterManager helixClusterManager =
+        new HelixClusterManager(clusterMapConfig, selfInstanceName, helixManagerFactory, new MetricRegistry());
 
     Counter initFailureCount = new Counter();
-    DynamicClusterChangeHandler dynamicChangeHandler =
-        new DynamicClusterChangeHandler(clusterMapConfig, localDc, selfInstanceName, Collections.emptyMap(),
-            mockManagerCallback, mockHandlerCallback,
-            new HelixClusterManagerMetrics(new MetricRegistry(), mockManagerCallback), e -> initFailureCount.inc(),
-            new AtomicLong());
+    HelixClusterChangeHandler helixClusterChangeHandler =
+        helixClusterManager.new HelixClusterChangeHandler(localDc, helixClusterManager.getManagerCallback(),
+            e -> initFailureCount.inc());
     // create an InstanceConfig with invalid entry that mocks error info added by Helix controller
     PartitionId selectedPartition = testPartitionLayout.getPartitionLayout().getPartitions(null).get(0);
     Replica testReplica = (Replica) selectedPartition.getReplicaIds().get(0);
@@ -267,15 +263,15 @@ public class ClusterChangeHandlerTest {
     InstanceConfigToDataNodeConfigAdapter.Converter converter =
         new InstanceConfigToDataNodeConfigAdapter.Converter(clusterMapConfig);
     // 1st call, to verify initialization code path
-    dynamicChangeHandler.onDataNodeConfigChange(Collections.singleton(converter.convert(instanceConfig)));
+    helixClusterChangeHandler.onDataNodeConfigChange(Collections.singleton(converter.convert(instanceConfig)));
     // 2nd call, to verify dynamic update code path
-    dynamicChangeHandler.onDataNodeConfigChange(Collections.singletonList(converter.convert(instanceConfig)));
+    helixClusterChangeHandler.onDataNodeConfigChange(Collections.singletonList(converter.convert(instanceConfig)));
     assertEquals("There shouldn't be initialization errors", 0, initFailureCount.getCount());
   }
 
   /**
    * Test new instances/partitions are added to cluster dynamically. {@link HelixClusterManager} with
-   * {@link DynamicClusterChangeHandler} should absorb the change and update in-mem cluster map.
+   * {@link HelixClusterChangeHandler} should absorb the change and update in-mem cluster map.
    * 1. add new instance
    * 2. add new partition onto new instance
    * 3. add new partition onto existing instance
@@ -531,7 +527,7 @@ public class ClusterChangeHandlerTest {
 
   /**
    * Test the case where current node receives InstanceConfig change triggered by itself due to replica addition. We need
-   * to verify {@link DynamicClusterChangeHandler} will check if new replica from InstanceConfig exists in bootstrap
+   * to verify {@link HelixClusterChangeHandler} will check if new replica from InstanceConfig exists in bootstrap
    * replica map. The intention here is to avoid creating a second instance of replica on current node.
    */
   @Test
@@ -644,13 +640,15 @@ public class ClusterChangeHandlerTest {
         new HelixClusterManager(clusterMapConfig, selfInstanceName, helixManagerFactory, new MetricRegistry());
     // disk capacity after initialization should equal to original value (100L * 1024 * 1024 * 1024)
     PartitionId partitionId = helixClusterManager.getAllPartitionIds(null).get(0);
-    assertEquals("Mismatch in disk capacity", 100L * 1024 * 1024 * 1024, partitionId.getReplicaIds().get(0).getDiskId().getRawCapacityInBytes());
+    assertEquals("Mismatch in disk capacity", 100L * 1024 * 1024 * 1024,
+        partitionId.getReplicaIds().get(0).getDiskId().getRawCapacityInBytes());
     // update disk capacity
     testHardwareLayout.updateDiskCapacity(500L * 1024 * 1024 * 1024);
     Utils.writeJsonObjectToFile(testHardwareLayout.getHardwareLayout().toJSONObject(), hardwareLayoutPath);
     helixCluster.upgradeWithNewHardwareLayout(hardwareLayoutPath);
     partitionId = helixClusterManager.getAllPartitionIds(null).get(0);
-    assertEquals("Mismatch in disk capacity", 500L * 1024 * 1024 * 1024, partitionId.getReplicaIds().get(0).getDiskId().getRawCapacityInBytes());
+    assertEquals("Mismatch in disk capacity", 500L * 1024 * 1024 * 1024,
+        partitionId.getReplicaIds().get(0).getDiskId().getRawCapacityInBytes());
     helixClusterManager.close();
   }
